# Add ProcessString to Pipeline.