--- a/pkg/chunkenc/memchunk.go
+++ b/pkg/chunkenc/memchunk.go
@@ -623,8 +623,7 @@ func (hb *headBlock) iterator(ctx context.Context, direction logproto.Direction,
 	streams := map[uint64]*logproto.Stream{}
 	for _, e := range hb.entries {
 		chunkStats.HeadChunkBytes += int64(len(e.s))
-		line := []byte(e.s)
-		newLine, parsedLbs, ok := pipeline.Process(line)
+		newLine, parsedLbs, ok := pipeline.ProcessString(e.s)
 		if !ok {
 			continue
 		}
@@ -638,7 +637,7 @@ func (hb *headBlock) iterator(ctx context.Context, direction logproto.Direction,
 		}
 		stream.Entries = append(stream.Entries, logproto.Entry{
 			Timestamp: time.Unix(0, e.t),
-			Line:      string(newLine),
+			Line:      newLine,
 		})
 
 	}
@@ -662,8 +661,7 @@ func (hb *headBlock) sampleIterator(ctx context.Context, mint, maxt int64, extra
 	series := map[uint64]*logproto.Series{}
 	for _, e := range hb.entries {
 		chunkStats.HeadChunkBytes += int64(len(e.s))
-		line := []byte(e.s)
-		value, parsedLabels, ok := extractor.Process(line)
+		value, parsedLabels, ok := extractor.ProcessString(e.s)
 		if !ok {
 			continue
 		}