--- a/pkg/ingester/tailer.go
+++ b/pkg/ingester/tailer.go
@@ -158,7 +158,7 @@ func (t *tailer) processStream(stream logproto.Stream) ([]logproto.Stream, error
 	}
 	sp := t.pipeline.ForStream(lbs)
 	for _, e := range stream.Entries {
-		newLine, parsedLbs, ok := sp.Process([]byte(e.Line))
+		newLine, parsedLbs, ok := sp.ProcessString(e.Line)
 		if !ok {
 			continue
 		}
@@ -171,7 +171,7 @@ func (t *tailer) processStream(stream logproto.Stream) ([]logproto.Stream, error
 		}
 		stream.Entries = append(stream.Entries, logproto.Entry{
 			Timestamp: e.Timestamp,
-			Line:      string(newLine),
+			Line:      newLine,
 		})
 	}
 	streamsResult := make([]logproto.Stream, 0, len(streams))